lib: Rename internal `oci` module to `ociwriter`
This makes fully clear this is just for writing; reading
is handled via the `containers-image-proxy` and `oci_spec` crates.
